requirements.txt
semver
flake8
black
toml
pyyaml